% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{add_packages}
\alias{add_packages}
\title{Adding packages}
\usage{
add_packages()
}
\value{
Nothing
}
\description{
Load the packages needed to run the agent-based CDI simulator
}
\examples{
add_packages();
}
