% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{basic_stats}
\alias{basic_stats}
\title{Basic Univariate Statistics of Data}
\usage{
basic_stats(data_in, key, samplesize, width, histo = TRUE, plot = FALSE)
}
\arguments{
\item{data_in}{Table of data to be analyzed}

\item{key}{Name of the variable to be analzyed}

\item{samplesize}{Number of random values from the data to be taken}

\item{histo}{TRUE if you want to plot a histogram, FALSE if you want to plot a bar graph}

\item{plot}{Decides whether data will be plotted or not}
}
\value{
A vector of random values taken from the data
}
\description{
Generate basic statistics for numerical data, as well as random values to be used in the agent-based simulation
}
\examples{
random_values <- basic_stats(data_in, key, samplesize, width, histo = TRUE, plot = FALSE)
}
