% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{CDI_predict_data}
\alias{CDI_predict_data}
\title{Reformat Data for Machine Learning}
\usage{
CDI_predict_data(master)
}
\arguments{
\item{master}{The data to be reformatted}
}
\value{
A new, reformatted version of the data
}
\description{
Remove some variables and convert some variable types to factors
}
\examples{
data <- CDI_predict_data(master);
}
