% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{contacts}
\alias{contacts}
\title{Contact Matrix}
\usage{
contacts(agents)
}
\arguments{
\item{agents}{The list of agents, both patients and healthcare workers, who will be in the simulation}
}
\value{
A symmetric matrix of size NxN, where N is the number of agents, containing number of contacts between each individual agent for each timestep
}
\description{
Randomly generate a symmetric matrix of contacts between agents
}
\examples{
contact_matrix <- contacts(agents)
}
