% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{dependent}
\alias{dependent}
\title{Patient-Level Dependent Variables}
\usage{
dependent(data, npat, age_pat, sex_pat)
}
\arguments{
\item{data}{Data to be used}

\item{npat}{Number of patients}

\item{age_pat}{Ages of patients}

\item{sex_pat}{Sexes of patients}
}
\value{
A tibble of patients with their individual characteristics
}
\description{
Draw dependent values for patients, ndx, elix, necode, orproc, npr, prior_visit, and hospitalunit, from patient data
}
\examples{
tibble_patients <- dependent(data, npat, age_pat, sex_pat)
}
