% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{diagnose}
\alias{diagnose}
\title{Machine-Learning Model Statistics}
\usage{
diagnose(model, model_type, train_data, test_data)
}
\arguments{
\item{model}{Machine-learning model}

\item{model_type}{Model type for plotting titles}

\item{train_data}{Data trained on by the model}

\item{test_data}{Data to test on by the model}
}
\value{
ROC curve and variable importance scores
}
\description{
Check model accuracy and statistics using ROC curves and variable importance
}
\examples{
importance <- diagnose(gbm_model, "GBM", train, test)
}
