% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{HCUP}
\alias{HCUP}
\title{Access Patient-Level Data}
\usage{
HCUP(
  link = "/home/ebarsotti/LSS/Statepi_Marketscan/databases/HCUP/IA.db",
  core = "IA_SID_2010_core"
)
}
\arguments{
\item{link}{The directory containing the patient data}

\item{core}{The table you would like to use from the patient data}
}
\value{
The patient database
}
\description{
Load patient data to be used in machine learning and generating patients for the simulator
}
\examples{
db <- HCUP(link = "/home/ebarsotti/LSS/Statepi_Marketscan/databases/HCUP/IA.db", core = "IA_SID_2010_core");
}
