% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{hcws_table}
\alias{hcws_table}
\title{Constructing Tibble of Healthcare Workers}
\usage{
hcws_table(tibble_pat, nhcw, data)
}
\arguments{
\item{tibble_pat}{Tibble of individual patients in the simulation}

\item{nhcw}{Number of healthcare workers in the simulation}

\item{data}{Patient-level data that is also used for finding out which hospital unit healthcare professionals work in}
}
\value{
A tibble of individual healthcare workers
}
\description{
Generate healthcare workers, of whom none are initially infected, with susceptibility equal to 0.05 quantile of patients and infectivity equal to 0.5 quantile of patients
}
\examples{
tibble_hcw <- hcws_table(tibble_pat, nhcw, data)
}
