% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{independent}
\alias{independent}
\title{Patient-Level Independent Variables}
\usage{
independent(data)
}
\arguments{
\item{data}{Data to be used}
}
\value{
A list of two vectors: age of patients and sex of patients
}
\description{
Draw independent values for patients, age and sex, from patient data
}
\examples{
list_of_age_and_sex <- independent(data)
}
