% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{init_agents}
\alias{init_agents}
\title{Initialize Agents}
\usage{
init_agents(
  nhcw = 800L,
  npat = 100L,
  hcw_feature_names,
  pat_feature_names,
  tibble_pat,
  tibble_hcw
)
}
\arguments{
\item{nhcw}{Number of healthcare workers in the simulation}

\item{npat}{Number of patients in the simulation}

\item{hcw_feature_names}{Individual healthcare worker feature names}

\item{pat_feature_names}{Individual patient feature names}

\item{tibble_pat}{Tibble of patients}

\item{tibble_hcw}{Tibble of healthcare workers}
}
\value{
A list containing two objects: the list of agents is the first; the initial set of infected agents is the second
}
\description{
Make one list of agents that contain both healthcare workers and patients, which is needed for the C++ disease propogation function
}
\examples{
all_agents <- init_agents(nhcw = 800L, npat = 100L, hcw_feature_names, pat_feature_names, tibble_pat, tibble_hcw)
}
