% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{install_packages}
\alias{install_packages}
\title{Installing Packages}
\usage{
install_packages()
}
\value{
Nothing
}
\description{
Install packages needed to run agent-based CDI simulator
}
\examples{
install_packages();
}
