% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{pat_los}
\alias{pat_los}
\title{Patient-Level Length of Stay Variable Generation}
\usage{
pat_los(data_one, data_two, npat)
}
\arguments{
\item{data_one}{Data for all patients}

\item{data_two}{Tibble of data only for patients to be used in this model}

\item{npat}{Number of patients in the tibble of patients}
}
\value{
An updated tibble of patients with a new length of stay column
}
\description{
Calculate length of stay for individual agents based on age, sex, OR procedure, and CDI status
}
\examples{
complete_tibble_patients <- pat_los(data_one, data_two, npat)
}
