% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{pat_preds}
\alias{pat_preds}
\title{Patient CDI Infection Probabilities}
\usage{
pat_preds(model, data, intial_infect_pat)
}
\arguments{
\item{model}{Machine-learning model to be used}

\item{data}{Tibble of patients}

\item{initial_infect_pat}{Initial number of patients to be set as infected}
}
\value{
A new tibble of patients with infectivities and susceptibilities based on probabilities from the machine-learning model
}
\description{
Calculate probabilities of patients being infected based on machine-learning algorithm
}
\examples{
new_tibble_patients <- pat_preds(model, data, initial_infect_pat)
}
