% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{patients_table}
\alias{patients_table}
\title{Constructing Tibble of Patients}
\usage{
patients_table(ml_data, all_data, npat, initial_infect_pat, model)
}
\arguments{
\item{ml_data}{Data to use in machine-learning algorithm}

\item{all_data}{All patient-level data available}

\item{npat}{Number of patients in the simulation}

\item{model}{Machine-learning model to use}

\item{initial_infected_pat}{Number of initial infected patients in the simulation}
}
\value{
A tibble of patients for use in the simuation; patients generated by multiple other functions in this library
}
\description{
Generate patients based on previous patient-level data on characteristics such as age, sex, Elixhauser risk score, number of diagnoses, etc.
}
\examples{
patients <- patients_table(ml_data, all_data, npat, initial_infect_pat, model)
}
