% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{quantiles}
\alias{quantiles}
\title{Convert Some Data to Quantiles}
\usage{
quantiles(data, columns, partition)
}
\arguments{
\item{data}{Data containing variables to be converted}

\item{columns}{Variable columns to be converted}

\item{partition}{Number of quantiles to use}
}
\value{
New version of the data
}
\description{
Convert some categorical variables to quantiles to reduce number of categories
}
\examples{
data <- quantiles(data, cols, 8);
}
