% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{simulate}
\alias{simulate}
\title{CDI Agent-Based Simulation Function}
\usage{
simulate(
  ml_data,
  mod,
  timestep = 24L,
  agents,
  contact_matrix,
  initial_set,
  plot_now = FALSE,
  num,
  secondary_infections,
  hcw_to_pat_infections,
  pat_to_hcw_infections,
  pat_to_pat_infections,
  hcw_to_hcw_infections,
  secondary_patient_infected,
  secondary_hcw_infected,
  R0,
  attack_curve,
  pat_to_hcw,
  hcw_to_pat,
  hcw_to_hcw,
  iters,
  discharged,
  total_pat,
  transmission_dir
)
}
\arguments{
\item{ml_data}{The data to be used in the machine learning algorithm to calculate risk of CDI}

\item{mod}{The machine-learning model to be used to calculate risk of CDI}

\item{timestep}{Number of iterations for the simulation to loop through}

\item{agents}{List of all agents in the simulation}

\item{contact_matrix}{Matrix of contacts between all individual agents}

\item{initial_set}{Initial set of infected agents}

\item{plot_now}{TRUE if you want to watch attack rate over time; FALSE if you don't}

\item{num}{Variable that keeps track of number of timesteps completed}

\item{secondary_infections}{Variable keeps track of number of secondary infections}

\item{hcw_to_pat_infections}{Variable keeps track of number of healthcare worker to patient transmissions}

\item{pat_to_hcw_infections}{Variable keeps track of number of patient to healthcare worker transmissions}

\item{pat_to_pat_infections}{Variable keeps track of number of patient to patient transmissions}

\item{hcw_to_hcw_infections}{Variable keeps track of number of healthcare worker to healthcare worker transmissions}

\item{secondary_patient_infected}{Variable keeps track of number of secondary infections spread to patients}

\item{secondary_hcw_infected}{Variable keeps track of number of secondary infections spread to healthcare workers}

\item{R0}{Variable to store the basic reproduction number of the model}

\item{attack_curve}{Variable to store the attack curve, currently just the total number of secondary infections, in the model}

\item{pat_to_hcw}{Variable to store number of transmissions from patients to healthcare workers as a vector for later use}

\item{hcw_to_pat}{Variable to store number of transmissions from healthcare workers to patients as a vector for later use}

\item{hcw_to_hcw}{Variable to store number of transmissions from healthcare workers to other healthcare workers as a list for later use}

\item{iters}{Variable to store iteration number as a list for later use}

\item{discharged}{Variable to store the number of discharged patients}

\item{total_pat}{Variable to store the total number of patients, initial plus discharged}

\item{transmission_dir}{Variable containing directory name where the C++ transmission function is stored}
}
\value{
Long list containing updated agents and simulation statistics
}
\description{
Take in the list of all agents and the contact matrix, call the C++ function "propogate" that spreads the infection, and return statistics about how much the infection spreads and to whom it spreads
}
\examples{
sim_results <- simulate(ml_data, mod, timestep = 24L, agents, contact_matrix, initial_set, plot_now = FALSE, num, secondary_infections, hcw_to_pat_infections, pat_to_hcw_infections, pat_to_pat_infections, hcw_to_hcw_infections, secondary_patient_infected, secondary_hcw_infected, R0, attack_curve, pat_to_hcw, hcw_to_pat, hcw_to_hcw, iters, discharged, total_pat, transmission_dir)
}
