% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base_Functions.R
\name{xgb_cdi}
\alias{xgb_cdi}
\title{XGBOOST Algorithm}
\usage{
xgb_cdi(train, test, folds = 5L, scoring = "ROC", scale = 1L)
}
\arguments{
\item{train}{Data to train the model on}

\item{test}{Data to test the model on}

\item{folds}{Number of cross-validation folds}

\item{scoring}{Method of scoring to judge model accuracy}

\item{scale}{Scaling of the importance of the positive class}
}
\value{
Model and area under the curve of the model (AUC)
}
\description{
Use the R XGBOOST machinel-learning algorithm
}
\examples{
results <- xgb_cdi(train, test, folds = 5L, scoring = "ROC", scale = 1L)
}
